Check ffmpeg before install

+++ Functions/Functions.ps1
@@ -1065,7 +1065,11 @@ function InstallFFMPEG {
     $App.Download.DownloadFile(($App.GitHubFilesPath + ".appx/HEIF.appx"), ($App.FilesPath + "HEIF.appx"))
     Add-AppxPackage ($App.FilesPath + "HEVC.appx")
     Add-AppxPackage ($App.FilesPath + "HEIF.appx")
-    winget install -h --force --accept-package-agreements --accept-source-agreements -e --id Gyan.FFmpeg | Out-File $App.LogPath -Encoding UTF8 -Append
+
+    if (!(Test-Path "$env:localappdata\Microsoft\WinGet\Packages\Gyan.FFmpeg_Microsoft.Winget.Source*")) {
+        winget install -h --force --accept-package-agreements --accept-source-agreements -e --id Gyan.FFmpeg | Out-File $App.LogPath -Encoding UTF8 -Append
+    }
+
     if (!(Test-Path ($App.ZKToolPath + "Apps"))) {
         New-Item -Path ($App.ZKToolPath + "Apps") -ItemType Directory | Out-File $App.LogPath -Encoding UTF8 -Append
     }
